meetingNoteTaker/backend/requirements.txt
openai-whisper
ollama
pytest
tqdm
uvicorn[standard]
fastapi
python-multipart